+++ src/main/java/es/curso/registro/repository/ProductRepository.java
@@ -12,5 +12,5 @@ import es.curso.registro.model.Producto;
 public interface ProductRepository extends JpaRepository<Producto, Integer>{
 
 	@Query("select p from Producto p where (p.nombre like %?1%) and (p.descripcion like %?2%) and (p.precio <?3)")
-	List<Producto> findProductByFiltro(String nombre, String descripcion, int precio);
+	List<Producto> findProductByFiltro(String nombre, String descripcion, double precio);
 }
