+++ src/main/java/es/curso/registro/service/ProductService.java
@@ -8,11 +8,11 @@ public interface ProductService {
 
 	List<Producto> getAll();
 
-	List<Producto> getProductByFiltro(String nombre, String descripcion, int precio);
+	List<Producto> getProductByFiltro(String nombre, String descripcion, double precio);
 
 	public void addProducto(Producto producto);
 
-	public void addProducto(String nombre, String descripcion, String marca, int precio, int cantidad);
+	public void addProducto(String nombre, String descripcion, String marca, double precio, int cantidad);
 
 	public Producto getProductById(Integer id);
